+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Matrix function `from_column_array`.
+
 - `Hash` implementation for matrices and affines.
 
 - `Default` implementation for matrices and affines.